+++ lib/screens/create/create_templates/longform.dart
@@ -18,8 +18,6 @@ import 'package:junto_beta_mobile/widgets/mentions/channel_search_list.dart';
 import 'package:junto_beta_mobile/widgets/mentions/mentions_search_list.dart';
 import 'package:provider/provider.dart';
 
-enum ListType { mention, channels, empty }
-
 class CreateLongform extends StatefulWidget {
   const CreateLongform({Key key, this.expressionContext, this.address})
       : super(key: key);
@@ -147,6 +145,14 @@ class CreateLongformState extends State<CreateLongform>
     _titleController.dispose();
     _titleFocus.dispose();
     _bodyFocus.dispose();
+  }
+
+  void toggleSearch(bool value) {
+    if (value != _showList) {
+      setState(() {
+        _showList = value;
+      });
+    }
   }
 
   @override
@@ -286,40 +292,13 @@ class CreateLongformState extends State<CreateLongform>
                                   });
                                 }
                               },
-                              onSuggestionVisibleChanged: (val) {
-                                if (val != _showList) {
-                                  setState(() {
-                                    _showList = val;
-                                  });
-                                }
-                              },
+                              onSuggestionVisibleChanged: toggleSearch,
                               hideSuggestionList: true,
-                              mentions: [
-                                Mention(
-                                  trigger: '@',
-                                  data: [...addedmentions, ...completeUserList],
-                                  style: TextStyle(
-                                    color: Theme.of(context).primaryColorDark,
-                                    fontWeight: FontWeight.w700,
-                                  ),
-                                  markupBuilder: (trigger, mention, value) {
-                                    return '[$trigger$value:$mention]';
-                                  },
-                                ),
-                                Mention(
-                                  trigger: '#',
-                                  disableMarkup: true,
-                                  data: [
-                                    ...addedChannels,
-                                    ...completeChannelsList
-                                  ],
-                                  style: TextStyle(
-                                    color: Theme.of(context).primaryColorDark,
-                                    fontWeight: FontWeight.w700,
-                                  ),
-                                  matchAll: true,
-                                ),
-                              ],
+                              mentions: getMention(
+                                context,
+                                [...addedmentions, ...completeUserList],
+                                [...addedChannels, ...completeChannelsList],
+                              ),
                             ),
                           ),
                           if (_showList &&
